[CLEANUP] add use case Without Global Database
